c/os: fix for drawin amd64
