+++ c/os/os_wasm.go
@@ -1,0 +1,1 @@
+package os
